% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_EffLength.R
\name{adjust_EffLength}
\alias{adjust_EffLength}
\title{adjust_EffLength}
\usage{
adjust_EffLength(
  ruvs,
 
    efflength_pathdir = "/Users/patterja/Box/Heiser_Atwater/SMMART_Normalization/DATA_matrices/efflengths.txt",
  includeBCCL = FALSE
)
}
\arguments{
\item{ruvs}{(data frame) rows are gene name}

\item{efflength_pathdir}{(string) efflengths.txt}
}
\value{
dataframe
}
\description{
Adjusts effective length measurements
}
