% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{bprle}
\alias{bprle}
\title{Prettier relative log expression (RLE) boxplots}
\usage{
bprle(mat, title = "RLE", ...)
}
\arguments{
\item{title}{(string): title of the box plot.}

\item{mat:}{make it matrix.}
}
\value{
boxplot
}
\description{
Prettier relative log expression (RLE) boxplots
}
