% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2hugo.R
\name{convert2hugo}
\alias{convert2hugo}
\title{convert ENSGids to HUGO gene names}
\usage{
convert2hugo(
  mat,
 
    targetid_file = "/Users/patterja/Box Sync/Heiser_SMMARTReport/REFERENCE_FILES/target_id.txt",
  combined = FALSE,
  protein_coding = FALSE
)
}
\arguments{
\item{mat}{(matrix) columns are sample names, rows are ENSG IDs}

\item{targetid_file}{(char string) path to target_id.txt}

\item{protein_coding}{(logical): filter for protein coding only and then aggregate to HUGO}
}
\value{
aggregated matrix with HUGO IDs
}
\description{
RNA matrix of ENSG identifiers converted and aggregated (summed) to HUGO
identifiers using targetid file from GENCODE v24.
}
\examples{
convert2hugo(mat)
}
\author{
Janice Patterson
}
