% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPredTest.R
\name{drugPredTest}
\alias{drugPredTest}
\title{DrugPrediction Test}
\usage{
drugPredTest(
  path_to_bcclsmmart,
  cell_line_response = "cell_line_response_threshold_0.50_large_and_small_screen.RData",
  state = "smmart_trained_machine_learning_model.RData",
  kbmtl_train = "kbmtl_semisupervised_classification_variational_train.R",
  kbmtl_test = "kbmtl_semisupervised_classification_variational_test.R",
  path_to_targetid = "target_id.txt",
  path_to_trusight = "Trusight_genes.csv",
  threshold = 0.5
)
}
\arguments{
\item{path_to_bcclsmmart}{(string):path to FINAL_norm.txt run through normalizeRUV in SMMARTFunctions.R}

\item{cell_line_response}{(string):cell_line_response_threshold_0.50_large_and_small_screen.RData}

\item{state}{(string): path to "smmart_trained_machine_learning_model.RData" from drugPredTrain}

\item{kbmtl_train}{(string): "kbmtl_semisupervised_classification_variational_train.R"}

\item{kbmtl_test}{(string): "kbmtl_semisupervised_classification_variational_test.R"}

\item{path_to_targetid}{(string): path to target_id.txt}

\item{path_to_trusight}{(string): path to gene list Trusight.csv}

\item{threshold}{(numeric): 0.25, 0.5, 0.75}
}
\value{
Y_predicted(data.frame): prediction
}
\description{
version 2
}
