% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugPredictionFunctions.R
\name{drugPredTrain}
\alias{drugPredTrain}
\title{DrugPrediction TRAINING}
\usage{
drugPredTrain(path_to_bcclsmmart,
  kbmtl_train = "kbmtl_semisupervised_classification_variational_train.R",
  kbmtl_test = "kbmtl_semisupervised_classification_variational_test.R",
  path_to_targetid = "target_id.txt",
  path_to_trusight = "Trusight_genes.csv", threshold = 0.5)
}
\arguments{
\item{path_to_bcclsmmart}{(string):path to FINAL_norm.txt run through normalizeRUV in SMMARTFunctions.R}

\item{kbmtl_train}{(string): "kbmtl_semisupervised_classification_variational_train.R"}

\item{kbmtl_test}{(string): "kbmtl_semisupervised_classification_variational_test.R"}

\item{path_to_targetid}{(string): path to target_id.txt}

\item{path_to_trusight}{(string): path to gene list Trusight.csv}

\item{threshold}{(numeric): 0.25, 0.5, 0.75}
}
\value{
smmart_trained_machine_learning_model.RData
}
\description{
DrugPrediction TRAINING
}
