% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{hmap}
\alias{hmap}
\title{Heatmap function}
\usage{
hmap(mat, title = "heatmap", ...)
}
\arguments{
\item{mat}{(matrix): RXC matrix with rownames and colnames}

\item{title}{(string): name of plot title}
}
\value{
heatmap plot
}
\description{
Wrapper around pheatmap
}
\examples{

Additional parameters

#color            = c("blue", "red"), 
#color            = c("#FFFFFF", colorRampPalette(rev(brewer.pal(9, "RdBu")))(200)),
#cluster_cols     = FALSE,
#cluster_rows     = FALSE,
#label_rows       = 
}
