% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeRUV.R
\name{normalizeRUV}
\alias{normalizeRUV}
\title{normalize RUV}
\usage{
normalizeRUV(
 
    exp_matx = read.csv("/Users/patterja/Box Sync/Heiser_Atwater/SMMART_NORMALIZATION/DATA_MATRICES/PROCESSED/tpm_breast_20190214.txt",
    sep = "\\t", row.names = 1),
  includeBCCL = FALSE,
  filtperc = 10,
  mincnt = 3,
  normtype = "NONE",
  setk
)
}
\arguments{
\item{filtperc}{(int): 10\% of SMMART samples will have counts greater than mincnt}

\item{mincnt}{(int): TPM use 3, counts to CPM use 10}

\item{normtype}{(string): "NONE", "CPM"}

\item{setk}{(numeric): number of factors to remove.
if includeBCCL=T use CPM, setk=7, filtperc=10, mincnt = 10,
if includeBCCL=F use TPM, setk=1, filtperc=10, mincnt = 3,}

\item{exp_matx:}{matrix of expression, rownames=genes}

\item{includeBCCL:}{TRUE OR FALSE}
}
\value{
list of matrices
}
\description{
Takes a matrix and filters by counts greater than 10 in filtperc \% of SMMART samples
}
