% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{pcaplt}
\alias{pcaplt}
\title{PCA plot function}
\usage{
pcaplt(mat, title = "PCA Plot")
}
\arguments{
\item{title}{(string): Title of PCA plot.}

\item{mat:}{dataframe, datatable. First column is gene names, rows are genes, columns are samples.}
}
\value{
pcaplot
}
\description{
PCA plot function
}
