% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{pcaplt_expset}
\alias{pcaplt_expset}
\title{PCA plot function with a metadata file and expression set}
\usage{
pcaplt_expset(mat, metadat, title = "PCA Plot")
}
\arguments{
\item{:}{mat (matrix): matrix}

\item{:}{metadat (matrix): rownames correspond to colnames of mat}

\item{:}{title (string): title of plot}
}
\description{
TODO: might be broken with metadata
}
