% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{pcaplt_grid}
\alias{pcaplt_grid}
\title{PCA plot of multiple principle components. 1vs2 2vs3 1vs3}
\usage{
pcaplt_grid(mat, md, title = "PCA Plots")
}
\arguments{
\item{md}{is the metadata, a row for each column is ideal.}

\item{title}{(string): title of the scatter.}

\item{mat:}{full matrix of expression}
}
\value{
pca plots
}
\description{
PCA plot of multiple principle components. 1vs2 2vs3 1vs3
}
