% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{pcaplt_grid_smmart}
\alias{pcaplt_grid_smmart}
\title{PCA plot of multiple principle components. 1vs2 2vs3 1vs3
Filtered by SMMART samples only}
\usage{
pcaplt_grid_smmart(mat, metadat = md, shape = "diagnosis",
  title = "PCA Plots")
}
\arguments{
\item{title}{(string): title of the scatter.}

\item{mat:}{full matrix of expression}

\item{md}{is the metadata, a row for each column is ideal.}
}
\value{
pca plots
}
\description{
PCA plot of multiple principle components. 1vs2 2vs3 1vs3
Filtered by SMMART samples only
}
