% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{pcaplt_grid_smmart_tissue}
\alias{pcaplt_grid_smmart_tissue}
\title{PCA plot of multiple PCs with tissue not batch as shape condition}
\usage{
pcaplt_grid_smmart_tissue(mat, metadat = md, shape = "tissue",
  title = "PCA Plots")
}
\arguments{
\item{title}{(string): title of the scatter}

\item{mat:}{matrix}

\item{md}{is the metadata, a row for each column is ideal}
}
\value{
pca plot
}
\description{
PCA plot of multiple PCs with tissue not batch as shape condition
}
