% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRUV.R
\name{processRUV}
\alias{processRUV}
\title{processRUV}
\usage{
processRUV(filtmat, includeBCCL = TRUE, setk)
}
\arguments{
\item{filtmat}{(matrix): filtered matrix of sample=columns, genes=rows}

\item{includeBCCL}{(logic): TRUE/FALSE}

\item{setk}{(numeric): number of factors to remove.
if includeBCCL=T use CPM + setk=7
if includeBCCL=F use TPM + setk=1}
}
\value{
ruvset
}
\description{
Filter by UHR reps by default. Filtmat must include UHR replicates
 For use in normalize RUV functions
}
