% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx2gene.R
\name{tx2gene}
\alias{tx2gene}
\title{tx2gene wrapper}
\usage{
tx2gene(
  path_to_target = "target_id.txt",
 
    path_to_kallisto_output = "/Volumes/exahead1/lustre1/HeiserLab/patterja/tatlow_SMMART/output"
)
}
\arguments{
\item{path_to_target}{(string): path to the target_id.txt. file that converts ENSEMBL ids to HGVS. Filtering by trusight genes}

\item{path_to_kallisto_output}{(string): path to output of kallisto}
}
\value{
genetxi
}
\description{
Convert kallisto output to gene level
}
