% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMMARTFunctions.R
\name{updateSMMART_datasets}
\alias{updateSMMART_datasets}
\title{updateSMMART_datasets}
\usage{

  updateSMMART_datasets(proj_output = "/Volumes/exahead1/lustre1/HeiserLab/patterja/tatlow_SMMART/output",
  target_filename = "/Users/patterja/Workspace/SMMART_HeiserNormalization/data/target_id.txt")
}
\arguments{
\item{proj_output}{(character string) full path to all kallisto output directory}

\item{target_filename}{(character string) full path to the target_id.txt file.}
}
\value{
matrix aggregated to gene gene counts.
}
\description{
TODO: needs updating from generate_smmart_predictions
}
