% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMMARTFunctions.R
\name{normalizeRUV}
\alias{normalizeRUV}
\title{normalize RUV}
\usage{

  normalizeRUV(pathtocnts = "/Users/patterja/Workspace/SMMART_HeiserNormalization/data/FINAL_whole.txt",
  includeBCCL = FALSE, filtperc = 10, normtype = "CPM")
}
\arguments{
\item{normtype}{(string): "NONE", "CPM"}

\item{pathtocnts:}{matrix of expression, rownames=genes}
}
\value{
list of matrices
}
\description{
Takes a matrix and filters by counts greater than 10 in filtperc % of SMMART samples
}
