% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMMARTFunctions.R
\name{processRUV}
\alias{processRUV}
\title{Filter by UHR reps by default. Filtmat must include UHR replicates}
\usage{
processRUV(filtmat, includeBCCL = FALSE)
}
\arguments{
\item{filtmat}{(matrix): filtered matrix of sample=columns, genes=rows}

\item{includeBCCL}{(logic): TRUE/FALSE}
}
\value{
ruvset
}
\description{
For use in normalize RUV functions
}
